requirements.txt
streamlit
pypdf
python-dotenv
requests
